Add Hertz waveguide test to integration tests
