+++ src/hephaestus_lib/hephaestus.hpp
@@ -1,3 +1,4 @@
+#pragma once
 #include "hephaestus_hertz.hpp"
 #include "hephaestus_joule.hpp"
