+++ src/hephaestus_lib/hertz_solver.hpp
@@ -8,10 +8,10 @@
 // MFEM is free software; you can redistribute it and/or modify it under the
 // terms of the GNU Lesser General Public License (as published by the Free
 // Software Foundation) version 2.1 dated February 1999.
+#pragma once
 
 #ifndef MFEM_HERTZ_SOLVER
 #define MFEM_HERTZ_SOLVER
-
 #include "../common/pfem_extras.hpp"
 #include "boundary_conditions.hpp"
