+++ src/hephaestus_lib/joule_solver.hpp
@@ -8,6 +8,7 @@
 // MFEM is free software; you can redistribute it and/or modify it under the
 // terms of the BSD-3 license. We welcome feedback and contributions, see file
 // CONTRIBUTING.md for details.
+#pragma once
 
 #ifndef MFEM_JOULE_SOLVER
 #define MFEM_JOULE_SOLVER
