+++ test/integration/CMakeLists.txt
@@ -1,15 +1,15 @@
 file(GLOB_RECURSE src_integration "*.cpp")
 
 add_executable(integration_tests ${src_integration})
-add_compile_options(integration_tests ${BUILD_TYPE_COMPILER_FLAGS})
+add_compile_options(integration_tests ${BUILD_TYPE_COMPILER_FLAGS} -fsanitize=address)
 target_include_directories(integration_tests PRIVATE ${GTEST_INCLUDE_DIRS})
 target_include_directories(integration_tests PRIVATE ${GMOCK_INCLUDE_DIRS})
-target_include_directories(integration_tests PUBLIC ${PROJECT_SOURCE_DIR}/src/lib)
+target_include_directories(integration_tests PUBLIC ${MFEM_COMMON_INCLUDES} ${MFEM_INCLUDE_DIRS})
+target_include_directories(integration_tests PUBLIC ${PROJECT_SOURCE_DIR}/src/${PROJECT_NAME}_lib)
 
 target_link_libraries(integration_tests ${GTEST_LIBRARY} ${GTEST_MAIN_LIBRARY} pthread)
 target_link_libraries(integration_tests ${CMAKE_LIBRARY_OUTPUT_DIRECTORY}/lib${PROJECT_NAME}.so)
-
-#install(TARGETS integration_tests DESTINATION bin)
+target_link_libraries(integration_tests ${MFEM_LIBRARIES} ${MFEM_COMMON_LIBRARY} -lrt)
 
 include(GoogleTest)
 gtest_discover_tests(integration_tests)
